adk-agent/requirements.txt
google-adk[extensions]==1.24.1
python-dotenv==1.2.1